Use redux-toolkit's createAsyncThunk and change the directory structure
