+++ zipkin-lens/src/components/App/App.jsx
@@ -22,7 +22,7 @@ import MomentUtils from '@date-io/moment';
 
 import Layout from './Layout';
 import DiscoverPage from '../DiscoverPage';
-import DependenciesPage from '../../features/dependencies/DependenciesPage';
+import DependenciesPage from '../../dependencies/DependenciesPage';
 import TracePage from '../TracePage';
 import { UiConfig, UiConfigConsumer } from '../UiConfig';
 import configureStore from '../../store/configure-store';
