+++ zipkin-lens/src/dependencies/DependenciesPage.test.jsx
@@ -21,7 +21,7 @@ import moment from 'moment';
 import React from 'react';
 
 import DependenciesPage from './DependenciesPage';
-import render from '../../test/util/render-with-default-settings';
+import render from '../test/util/render-with-default-settings';
 
 jest.mock('@material-ui/pickers', () => {
   // eslint-disable-next-line global-require
